requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scipy
tqdm
opencv-python
progress
pillow